Move client telemetry logic into the HttpClient
