+++ stripe/http_client.py
@@ -7,9 +7,11 @@ import email
 import time
 import random
 import threading
+import json
 
 import stripe
 from stripe import error, util, six
+from stripe.request_metrics import RequestMetrics
 
 # - Requests is the preferred HTTP library
 # - Google App Engine has urlfetch
@@ -61,6 +63,10 @@ except ImportError:
 from stripe.six.moves.urllib.parse import urlparse
 
 
+def _now_ms():
+    return int(round(time.time() * 1000))
+
+
 def new_default_http_client(*args, **kwargs):
     if urlfetch:
         impl = UrlFetchClient
@@ -104,10 +110,19 @@ class HTTPClient(object):
 
         self._thread_local = threading.local()
 
+        self._last_request_metrics = None
+
     def request_with_retries(self, method, url, headers, post_data=None):
+        if stripe.enable_telemetry and self._last_request_metrics:
+            headers["X-Stripe-Client-Telemetry"] = json.dumps(
+                {"last_request_metrics": self._last_request_metrics.payload()}
+            )
+
         num_retries = 0
 
         while True:
+            request_start = _now_ms()
+
             try:
                 num_retries += 1
                 response = self.request(method, url, headers, post_data)
@@ -134,6 +149,14 @@ class HTTPClient(object):
                 time.sleep(sleep_time)
             else:
                 if response is not None:
+                    rheaders = response[2]
+                    if "Request-Id" in rheaders and stripe.enable_telemetry:
+                        request_id = rheaders["Request-Id"]
+                        request_duration_ms = _now_ms() - request_start
+                        self._last_request_metrics = RequestMetrics(
+                            request_id, request_duration_ms
+                        )
+
                     return response
                 else:
                     raise connection_error
